Extract dashboard panel order helpers
